+++ pages/api/projects/[id]/index.ts
@@ -1,5 +1,6 @@
 import dbConnect from "../../../../utils/dbConnect";
 import Project from "../../../../models/projects";
+import { getSession } from "next-auth/react";
 
 // Connect database
 dbConnect();
@@ -9,6 +10,8 @@ export default async (req, res) => {
     query: { id },
     method,
   } = req;
+
+  const session = await getSession({ req });
 
   switch (method) {
     case "GET":
@@ -25,31 +28,39 @@ export default async (req, res) => {
       }
       break;
     case "PUT":
-      try {
-        const data = await Project.findByIdAndUpdate(id, req.body, {
-          new: true,
-          runValidators: true,
-        });
-        if (!data) {
-          return res.status(400).json({ success: false });
-        }
+      if (!session) {
+        res.status(401).json({ error: "Unauthenticated User" });
+      } else {
+        try {
+          const data = await Project.findByIdAndUpdate(id, req.body, {
+            new: true,
+            runValidators: true,
+          });
+          if (!data) {
+            return res.status(400).json({ success: false });
+          }
 
-        res.status(200).json({ success: true, data });
-      } catch (error) {
-        res.status(400).json({ success: false });
+          res.status(200).json({ success: true, data });
+        } catch (error) {
+          res.status(400).json({ success: false });
+        }
       }
       break;
     case "DELETE":
-      try {
-        const deletedProject = await Project.deleteOne({ _id: id });
+      if (!session) {
+        res.status(401).json({ error: "Unauthenticated User" });
+      } else {
+        try {
+          const deletedProject = await Project.deleteOne({ _id: id });
 
-        if (!deletedProject) {
-          return res.status(400).json({ success: false });
+          if (!deletedProject) {
+            return res.status(400).json({ success: false });
+          }
+
+          res.status(200).json({ success: true, data: {} });
+        } catch (error) {
+          res.status(400).json({ success: false });
         }
-
-        res.status(200).json({ success: true, data: {} });
-      } catch (error) {
-        res.status(400).json({ success: false });
       }
       break;
